refactor(services): convert sponsorship service to pure factory function
